--- a/src/histogram.c
+++ b/src/histogram.c
@@ -205,9 +205,9 @@ static void his_render(void *data, gs_effect_t *effect)
 		uint8_t *video_data = NULL;
 		uint32_t video_linesize;
 		PROFILE_START(prof_draw_histogram_name);
-		if (gs_stagesurface_map(src->cm.stagesurface, &video_data, &video_linesize)) {
+		if (cm_stagesurface_map(&src->cm, &video_data, &video_linesize)) {
 			his_draw_histogram(src, video_data, video_linesize);
-			gs_stagesurface_unmap(src->cm.stagesurface);
+			cm_stagesurface_unmap(&src->cm);
 		}
 		PROFILE_END(prof_draw_histogram_name);
 	}