… CMakeLists.txt
… README.md
… doc/dock.md
… src/roi.c
… src/roi.h
… src/scope-dock.cpp
… src/scope-dock.hpp
… src/scope-widget-properties.cpp
… src/scope-widget.cpp
… src/scope-widget.hpp
… src/vectorscope.c
… src/waveform.c