This is a new feature.
User can drag on ROI in dock and the  region will be calculated in vectorscope, waveform, and histogram.
In addition, stage-surface is called in ROI once and other sources can re-use it so that it will improve rendering time.
![screenshot-roi-example](https://user-images.githubusercontent.com/780600/125464622-0fd6ad92-15b1-4fca-b723-b59a052ea7bd.png)

<!-- If unsure, feel free to let them unchecked. -->
- [x] The commit is reviewed by yourself.
- [x] The code is tested.
- [x] Document is up to date or not necessary to be changed.
- [x] The commit is compatible with repository's license.

#### Performance
This PR will improve rendering time in some GPU.
Before this PR:  `render_displays: min=0.007 ms, median=50.246 ms, max=105.506 ms, 99th percentile=52.425 ms`
With this PR: `render_displays: min=0.007 ms, median=25.902 ms, max=202.326 ms, 99th percentile=33.499 ms`
Environment of above are measured on CentOS 8 with AMD RV620 and Core2 Duo E7200. This is because number of transferring textures from GPU to CPU is reduced from 3 per frame to 1 per frame. I'd like to note that the CPU and GPU are very old.